Use String instead of &str in Report, Annotation structs
This change improves the ergonomics of the library, and since Code
Insights reports are by nature created and then sent over the wire as a
one-shot process (as opposed to sticking around in memory for any
significant amount of time), bending over backwards to avoid Report or
Annotation owning the data they contain doesn't seem worth the trouble.

Also use separate structs for the builders, since that seems to be the
most common pattern based on looking at the code of some highly regarded
Rust projects.